Use BCD to collect specs referenced by web-features entries
The `spec` property in web-features is rather coarse. BCD keys are more
specific. The code now goes through the BCD keys targeted a feature to
collect the list of specs that define part of the feature under consideration.

When BCD references a spec without specifying its version, the code maps to the
current specification in the series.

Important note: to be able to use the latest version of web-features, the
dependency to `web-features` is now to a local folder.